+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Added
 -----
 * Make training scheduler and batch_size configurable.
 * Enable multiple active data versions on Study objects.
+* Can now take notes in the BaseStudy `notes` field.
 
 Fixed
 -----